dom/tests/mochitest/bugs/test_devicemotion_multiple_listeners.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=673472
-->
<head>
  <title>Test for Bug 673472</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <base href="http://mochi.test:8888/tests/dom/tests/mochitest/bugs/">
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=673472">Mozilla Bug 673472</a>
<pre id="test">
<script type="application/javascript">

function f() {
try {
  var ifr = document.getElementById('i');
  if (ifr.contentWindow.document.location == 'about:blank') {
    ifr.parentNode.removeChild(ifr);
    ok(true, "no crash received");
    SpecialPowers.executeSoon(SimpleTest.finish);
  }
} catch(e) {
ok(false, e);
}
}

SimpleTest.waitForExplicitFinish();
</script>
</pre>
<iframe src="devicemotion_outer.html" onload="f()" id=i></iframe>  
</body>
</html>
